Add files via upload
Made an Arduino script which uses a Madgwick filter to determine the roll and pitch, then uses the magnetometer to determine the heading. From this, the accelerations relative to the global frame are determined using a rotation matrix. The global acceleration is double integrated to give the relative displacement from the starting position. There is a large amount of noise however, a filter would increase the accuracy.